able to resize screen
the problem is that if you uses .initscr() you have to manually detect the  if the terminal was been resize and resize the screen to the terminal screen.
But using wrapper(scr) it will done it automatically and do not need to worry about the screen update I just need to update the maxY, maxX and resize the subscreen which is on the screen.

Problem :
Able to resize screen to bigger size but when screen shrinks program will end it self. Probably some error  occur in the process causing it to end the program.